server/requirements.txt
requests
beautifulsoup4
fastapi[all]
python-dotenv
openai
